+++ trunk/code/kmeans_omp.h
@@ -32,12 +32,12 @@ inline T array_load(const T *arr, const int main_iter, const int sub_iter, const
 }
 */
 
-int kmeans_omp(uchar *data, float *centers, int particle_count, int dimensions, int cluster_count, uchar *assignments, int grainSize);
+int kmeans_omp(uchar *data, double *centers, int particle_count, int dimensions, int cluster_count, uchar *assignments, int grainSize);
 
-// int select_centers_omp(uchar *data, int particle_count, int dimensions, int cluster_count, float *centers);
-int select_centerspp_omp(uchar *data, int particle_count, int dimensions, int cluster_count, float *centers);
+// int select_centers_omp(uchar *data, int particle_count, int dimensions, int cluster_count, double *centers);
+int select_centerspp_omp(uchar *data, int particle_count, int dimensions, int cluster_count, double *centers);
 
-// float compute_distance(uchar *particle_data, const int particle_iter, float *centers, const int center_iter, const int dimensions);
+// double compute_distance(uchar *particle_data, const int particle_iter, double *centers, const int center_iter, const int dimensions);
 
 #endif
